Started adapting get_string to be able to... ...implement the algorythm laid on algorithm.txt.
This means get_string is being divided into get_row_string and get_col_string
and each of their auxs are being refactored to have a static call
counter and a (prev) tag to be able to know that it should return.
get_row_string already passed several tests, but we'll test more.

	new file:   get_row_string.c
	deleted:    get_string.c
	modified:   include.h